Disable cursor and imgui
ImGui sets the cursor mode to be enabled every frame. The way the cs488
framework is written makes it difficult to disable the cursor, which is
something that this game must do. For now we can disable ImGui, but when I work
on the main menu I'll probably need to go back and rewrite CS488Window::run or
hack some state machine together that lets me programmatically toggle ImGui.