Update ImGui to latest release
I think it was able to successfully port over the framework code to leverage the
newest release of ImGui. I may need to refactor the framework code to get it to
play well with the main menu.

I'm putting this in one big commit in case I need to revert at some point.